+++ src/main/kotlin/one/oktw/galaxy/item/type/MaterialType.kt
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package one.oktw.galaxy.item.enums
+package one.oktw.galaxy.item.type
 
 enum class MaterialType(val customModelData: Int) {
     DUMMY(0)
